+++ pvr/src/cube/ICube.java
@@ -2,9 +2,9 @@ package cube;
 
 public interface ICube {
 
-	public double getCurrentTemp();
-	public void setInitTemp(double temp);
-	public void setNewTemp(double temp);
+	public float getCurrentTemp();
+	public void setInitTemp(float temp);
+	public void setNewTemp(float temp);
 	public void setNewTemp(ICube[][][] block, int x, int y, int z);
 	public void updateTemp();
 }
